+++ lib/ui/groups/GroupDetailPanel.dart
@@ -91,11 +91,12 @@ class GroupDetailPanel extends StatefulWidget with AnalyticsInfo {
 
   final Group? group;
   final String? groupIdentifier;
-  final String? groupPostId;
-  final String? groupPostCommentId;
+  final String? groupPostId; //Preload post
+  final String? groupPostCommentId; //Preload post comment
+  final String? groupEventId; //Preload event
   final AnalyticsFeature? _analyticsFeature;
 
-  GroupDetailPanel({this.group, this.groupIdentifier, this.groupPostId, AnalyticsFeature? analyticsFeature, this.groupPostCommentId}) :
+  GroupDetailPanel({this.group, this.groupIdentifier, this.groupPostId, AnalyticsFeature? analyticsFeature, this.groupPostCommentId, this.groupEventId}) :
     _analyticsFeature = analyticsFeature;
 
   @override
@@ -410,7 +411,7 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> with NotificationsL
         //   _currentTab = DetailTab.About; //TBD
         // }
         _tabs = _buildDetailTabs();
-        _redirectToGroupPostIfExists();
+        _redirectToInnerDetailIfNeeded();
         _loadGroupAdmins();
 
         _updateController.add(GroupDetailPanel.notifyRefresh);
@@ -440,7 +441,7 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> with NotificationsL
   ///
   /// Loads group post by id (if exists) and redirects to Post detail panel
   ///
-  void _redirectToGroupPostIfExists() {
+  void _redirectToInnerDetailIfNeeded() {
     if ((_groupId != null) && (_postId != null)) {
       _increaseProgress();
       Social().loadSinglePost(groupId: _group!.id, postId: _postId!).then((post) {
@@ -450,6 +451,8 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> with NotificationsL
           Navigator.push(context, CupertinoPageRoute(builder: (context) => GroupPostDetailPanel(group: _group!, post: post, visibleCommentId: widget.groupPostCommentId, analyticsFeature: widget.analyticsFeature)));
         }
       });
+    } else if (_groupId != null && StringUtils.isNotEmpty(widget.groupEventId)){
+        Navigator.push(context, CupertinoPageRoute(builder: (context) => Event2DetailPanel(group: _group!, eventId: widget.groupEventId,)));
     }
   }