+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Show list items in Assistant response [#5334](https://github.com/rokwire/illinois-app/issues/5334).
 ### Fixed
 - Hide link to floorplans if the building has no floors in the campus data [#5340](https://github.com/rokwire/illinois-app/issues/5340).
+- Ensure that the content in Customize panel is visible [#5354](https://github.com/rokwire/illinois-app/issues/5354).
 
 ## [7.3.2] - 2025-08-13
 ### Fixed